Move TagClient.ts to types folder
