+++ ui/src/Locations/LocationClient.ts
@@ -17,7 +17,7 @@
 
 import Axios, {AxiosResponse} from 'axios';
 import {TagRequest} from '../Types/TagRequest';
-import {TagClient} from '../Tags/TagClient.interface';
+import {TagClient} from '../Types/TagClient';
 import {getToken} from '../Auth/TokenProvider';
 import {Space} from 'Types/Space';
 import MatomoEvents from '../Matomo/MatomoEvents';
