+++ ui/src/ReusableComponents/FormTagsField.tsx
@@ -23,7 +23,7 @@ import {AxiosResponse} from 'axios';
 import {useOnLoad} from './UseOnLoad';
 import {TagRequest} from '../Types/TagRequest';
 import SelectWithCreateOption, {Metadata} from '../ModalFormComponents/SelectWithCreateOption';
-import {TagClient} from '../Tags/TagClient.interface';
+import {TagClient} from '../Types/TagClient';
 import {useRecoilValue} from 'recoil';
 import {CurrentSpaceState, UUIDForCurrentSpaceSelector} from '../State/CurrentSpaceState';
